% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_arrow_c_stream_interface.R
\name{as_array_stream}
\alias{as_array_stream}
\title{Convert To 'Arrow C Stream Interface'.}
\usage{
as_array_stream(x, ..)
}
\arguments{
\item{x}{An R object to convert to an Arrow C stream interface.}
}
\description{
Convert To 'Arrow C Stream Interface'.
}
\seealso{
\url{https://arrow.apache.org/docs/format/CStreamInterface.html}
}
